+++ both/models/Issue.jsx
@@ -1,7 +1,7 @@
 Issues = [
   {
     name:"Toilet won't flush",
-    description:"Toilet won't flush. Brown liquid leaking onto floor.",
+    description:"Toilet won't flush. Brown liquid leaking onto floor. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. ",
     status:"New",
     urgent:true,
     thumb:1,
@@ -11,16 +11,22 @@ Issues = [
     contact:{
       name:"John Smith",
       email:"johnny@flart.flart",
-      phone:"0444-123-321"
+      phone:"0444-123-321",
+      thumb:"a1.jpg"
     },
-    symbol:"Hg",
+    supplier:{
+      name:"XYY Contractors",
+      email:"xyz@abc.123",
+      phone:"0400-123-123",
+      thumb:"supplier-1.png"
+    },
     number:80,
     weight:200.59,
     tags:"transition metal"
   },
   {
     name:"Rust from ceiling fan",
-    description:"Ceiling fan dropping rust.",
+    description:"Ceiling fan dropping rust. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.",
     status:"New",
     thumb:2,
     facility:{
@@ -29,16 +35,22 @@ Issues = [
     contact:{
       name:"John Smith",
       email:"johnny@flart.flart",
-      phone:"0444-123-321"
+      phone:"0444-123-321",
+      thumb:"a2.jpg"
     },
-    symbol:"Te",
+    supplier:{
+      name:"XYY Contractors",
+      email:"xyz@abc.123",
+      phone:"0400-123-123",
+      thumb:"supplier-2.png"
+    },
     number:52,
     weight:127.6,
     tags:"metalloid"
   },
   {
     name:"Shocking doors",
-    description:"Customers entering through sliding doors receiving electric shocks.",
+    description:"Customers entering through sliding doors receiving electric shocks. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.",
     status:"New",
     urgent:true,
     thumb:3,
@@ -48,16 +60,22 @@ Issues = [
     contact:{
       name:"John Smith",
       email:"johnny@flart.flart",
-      phone:"0444-123-321"
+      phone:"0444-123-321",
+      thumb:"a3.jpg"
     },
-    symbol:"Bi",
+    supplier:{
+      name:"XYY Contractors",
+      email:"xyz@abc.123",
+      phone:"0400-123-123",
+      thumb:"supplier-3.png"
+    },
     number:83,
     weight:208.980,
     tags:"post-transition metal"
   },
   {
     name:"Barbed wire on floor",
-    description:"Barbed wire in staff room a tripping hazard.",
+    description:"Barbed wire in staff room a tripping hazard. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.",
     status:"Open",
     thumb:4,
     facility:{
@@ -66,9 +84,15 @@ Issues = [
     contact:{
       name:"John Smith",
       email:"johnny@flart.flart",
-      phone:"0444-123-321"
+      phone:"0444-123-321",
+      thumb:"a4.jpg"
     },
-    symbol:"Pb",
+    supplier:{
+      name:"XYY Contractors",
+      email:"xyz@abc.123",
+      phone:"0400-123-123",
+      thumb:"supplier-4.png"
+    },
     number:82,
     weight:207.2,
     tags:"post-transition metal"
@@ -84,16 +108,22 @@ Issues = [
     contact:{
       name:"John Smith",
       email:"johnny@flart.flart",
-      phone:"0444-123-321"
+      phone:"0444-123-321",
+      thumb:"a5.jpg"
     },
-    symbol:"Au",
+    supplier:{
+      name:"XYY Contractors",
+      email:"xyz@abc.123",
+      phone:"0400-123-123",
+      thumb:"supplier-1.png"
+    },
     number:79,
     weight:196.967,
     tags:"transition metal"
   },
   {
     name:"Exploding fire extinguishers",
-    description:"Fire extinguishers exploding at random intervals.",
+    description:"Fire extinguishers exploding at random intervals. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.",
     status:"New",
     thumb:1,
     facility:{
@@ -102,16 +132,22 @@ Issues = [
     contact:{
       name:"John Smith",
       email:"johnny@flart.flart",
-      phone:"0444-123-321"
+      phone:"0444-123-321",
+      thumb:"a6.jpg"
     },
-    symbol:"K",
+    supplier:{
+      name:"XYY Contractors",
+      email:"xyz@abc.123",
+      phone:"0400-123-123",
+      thumb:"supplier-2.png"
+    },
     number:19,
     weight:39.0983,
     tags:"alkali metal"
   },
   {
     name:"Cracked walls, level 7",
-    description:"Large cracks appearing in walls on level 7.",
+    description:"Large cracks appearing in walls on level 7. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.",
     status:"Open",
     urgent:true,
     thumb:2,
@@ -121,9 +157,15 @@ Issues = [
     contact:{
       name:"John Smith",
       email:"johnny@flart.flart",
-      phone:"0444-123-321"
+      phone:"0444-123-321",
+      thumb:"a7.jpg"
     },
-    symbol:"Na",
+    supplier:{
+      name:"Arvis A",
+      email:"xyz@abc.123",
+      phone:"0400-123-123",
+      thumb:"supplier-3.png"
+    },
     number:11,
     weight:39.0983,
     tags:"alkali metal"
@@ -139,16 +181,22 @@ Issues = [
     contact:{
       name:"John Smith",
       email:"johnny@flart.flart",
-      phone:"0444-123-321"
+      phone:"0444-123-321",
+      thumb:"a8.jpg"
     },
-    symbol:"Cd",
+    supplier:{
+      name:"XYY Contractors",
+      email:"xyz@abc.123",
+      phone:"0400-123-123",
+      thumb:"supplier-4.png"
+    },
     number:48,
     weight:112.411,
     tags:"transition metal"
   },
   {
     name:"Nixious gas from AC",
-    description:"Air conditioner producing noxious gas.",
+    description:"Air conditioner producing noxious gas. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.",
     status:"Open",
     thumb:4,
     facility:{
@@ -157,16 +205,22 @@ Issues = [
     contact:{
       name:"John Smith",
       email:"johnny@flart.flart",
-      phone:"0444-123-321"
+      phone:"0444-123-321",
+      thumb:"a1.jpg"
     },
-    symbol:"Ca",
+    supplier:{
+      name:"QQQ Maintenence",
+      email:"xyz@abc.123",
+      phone:"0400-123-123",
+      thumb:"supplier-1.png"
+    },
     number:20,
     weight:40.078,
     tags:"alkaline-earth metal"
   },
   {
     name:"Non-ergonomic chairs",
-    description:"Chairs on level 6 are not ergonomic.",
+    description:"Chairs on level 6 are not ergonomic. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.",
     status:"Issued",
     thumb:5,
     facility:{
@@ -175,16 +229,22 @@ Issues = [
     contact:{
       name:"John Smith",
       email:"johnny@flart.flart",
-      phone:"0444-123-321"
+      phone:"0444-123-321",
+      thumb:"a2.jpg"
     },
-    symbol:"Re",
+    supplier:{
+      name:"XYY Contractors",
+      email:"xyz@abc.123",
+      phone:"0400-123-123",
+      thumb:"supplier-2.png"
+    },
     number:75,
     weight:186.207,
     tags:"transition metal"
   },
   {
     name:"Intermittent muzak",
-    description:"Elevator muzak intermittently cuts out.",
+    description:"Elevator muzak intermittently cuts out. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.",
     status:"New",
     thumb:1,
     facility:{
@@ -193,16 +253,22 @@ Issues = [
     contact:{
       name:"John Smith",
       email:"johnny@flart.flart",
-      phone:"0444-123-321"
+      phone:"0444-123-321",
+      thumb:"a3.jpg"
     },
-    symbol:"Tl",
+    supplier:{
+      name:"XYY Suppliers",
+      email:"xyz@abc.123",
+      phone:"0400-123-123",
+      thumb:"supplier-3.png"
+    },
     number:81,
     weight:204.383,
     tags:"post-transition metal"
   },
   {
     name:"Coffee machine leaking",
-    description:"Coffee machine producing putrid green liquid.",
+    description:"Coffee machine producing putrid green liquid. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.",
     status:"New",
     thumb:2,
     facility:{
@@ -211,16 +277,22 @@ Issues = [
     contact:{
       name:"John Smith",
       email:"johnny@flart.flart",
-      phone:"0444-123-321"
+      phone:"0444-123-321",
+      thumb:"a4.jpg"
     },
-    symbol:"Sb",
+    supplier:{
+      name:"XYY Contractors",
+      email:"xyz@abc.123",
+      phone:"0400-123-123",
+      thumb:"supplier-4.png"
+    },
     number:51,
     weight:121.76,
     tags:"metalloid"
   },
   {
     name:"Sparks from light switches",
-    description:"Sparks and smoke regularly emitted from light switches.",
+    description:"Sparks and smoke regularly emitted from light switches. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.",
     status:"Closed",
     thumb:3,
     facility:{
@@ -229,16 +301,22 @@ Issues = [
     contact:{
       name:"John Smith",
       email:"johnny@flart.flart",
-      phone:"0444-123-321"
+      phone:"0444-123-321",
+      thumb:"a5.jpg"
     },
-    symbol:"Co",
+    supplier:{
+      name:"ABC Contractors",
+      email:"xyz@abc.123",
+      phone:"0400-123-123",
+      thumb:"supplier-1.png"
+    },
     number:27,
     weight:58.933,
     tags:"transition metal"
   },
   {
     name:"Un-funny jokes",
-    description:"Visitor mugs in staff room printed with un-funny joke.",
+    description:"Visitor mugs in staff room printed with un-funny joke. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.",
     status:"Issued",
     thumb:4,
     facility:{
@@ -247,9 +325,15 @@ Issues = [
     contact:{
       name:"John Smith",
       email:"johnny@flart.flart",
-      phone:"0444-123-321"
+      phone:"0444-123-321",
+      thumb:"a6.jpg"
     },
-    symbol:"Yb",
+    supplier:{
+      name:"DEF Contractors",
+      email:"xyz@abc.123",
+      phone:"0400-123-123",
+      thumb:"supplier-2.png"
+    },
     number:70,
     weight:173.054,
     tags:"lanthanoid metal"