+++ both/views/layouts/LayoutMain.jsx
@@ -51,11 +51,11 @@ MainLayout = React.createClass({
 
     // FIXED-SIDEBAR
     // Uncomment this if you want to have fixed left navigation
-    /*$('body').addClass('fixed-sidebar');
+    $('body').addClass('fixed-sidebar');
      $('.sidebar-collapse').slimScroll({
          height: '100%',
          railOpacity: 0.9
-     });*/
+     });
     
     // slim scroll for right sidebar
     $('.sidebar-container').slimScroll({
@@ -78,7 +78,7 @@ MainLayout = React.createClass({
           <div id="pattern-bg" className="pattern-bg"></div>
           <TopNavBar />
           <main>{this.props.content}</main>
-          <Footer />
+          {/*<Footer />*/}
         </div>
         <RightSideBar />
       </div>